kvm-test: add option to add extra disks for install
The patch introduces a new option --extra-disk-size to create additional
disks in kvm-test. The option can be specified multiple times and will
not affect the first disk (the one boots in presence of the --boot
option).

 $ kvm-test.py --install --extra-disk-size 10G

... will create the first disk normally and will add an extra 10GiB disk

 $ kvm-test.py --extra-disk-size 10G --extra-disk-size 20G

... will create the first disk normally and will two extra disks (one of
10 and one of 20 GiB).

The size of the first disk (that can be honor the --boot option) can
still be configured using the --disksize option).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>