storage: fix crash when editing encrypted VG created in guided storage
When editing an encrypted VG that was created in the guided storage
screen, the VG information is originating from the server. However, the
server does not send the LUKS key over the wire. Instead it sends the
path to a keyfile which contains the key. The client may or may not have
read access to this keyfile so it does not have a reliable way to
determine the key.

This causes problem when editing the VG because the GUI expects to
receive a key when encryption is enabled.

If the VG object only contains a keyfile, the passphrase is set to None
and this result in the GUI crashing.

This patch fixes the crash by passing an empty passphrase instead of a
None value when the VG object only contains a keyfile.

This means the user gets forced to supply a passphrase again when
editing an encrypted VG that was created in the guided partition screen.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit dcc66fa346e1d8c6e484507f38bc9f1d3860a5f0)