apt: set cdrom APT source in etc/apt/sources.list.d instead of sources.list
To set the cdrom as an available APT source when installing the system,
we used to write to etc/apt/sources.list in the installed tree.
Then to revert the change in the installed system, we used to copy the
etc/apt/sources.list from the configured overlay to the installed
system, effectively discarding our change.

Nowadays though, the use of etc/apt/sources.list is discouraged and its
presence optional. When reverting, this means:
 * deleting etc/apt/sources.list from the installed system if the file
   was created by subiquity
 * otherwise, copying etc/apt/sources.list from the configured tree to
   the installed system

Rather than implementing the above logic, we stop using
etc/apt/sources.list anymore and instead write to
etc/apt/sources.list.d/01sy-cdrom.list.

This makes the deconfigure() method simpler because we all we need to
do is to erase the 01sy-cdrom.list file from the target system.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>