storage: accept to specify the source-id
Internally, SystemGetter.get(variation) expects that the variation exists for
the currently configured source.

However, because of the lock we added to SystemGetter.get(variation), it is now
more likely that the configured source has changed by the time we start
mounting stuff and querying snapd.

When that happens, SystemGetter.get(variation) fails with KeyError ; because
the variation is no longer present in the currently configured source.

Fixed by making sure that we use the source handler that was configured when
SystemGetter.get(variation) was first invoked.

LP: #2084032

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>